feat: 🌓 extract theme effect logic and improve theme handling
